code: move admin-locatipns-new to its own file
