+++ src/components/admin-locations-new/styles.module.scss
@@ -1,0 +1,32 @@
+.appBarWrapper { background-color: #fff; }
+
+.moduleContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 64px;
+  overflow-y: auto;
+  height: 100%;
+}
+.moduleWrapper {
+  margin-bottom: 24px;
+  padding-left: 16px;
+  padding-right: 16px;
+  width: 100%;
+  max-width: 848px;
+  box-sizing: border-box;
+}
+
+.arrow {
+  cursor: pointer;
+  margin-right: 16px;
+  &:hover { opacity: 0.9; }
+  &:active { opacity: 1; }
+}
+
+.centered {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
